Move kapacitor event to event-engine-common
