--- src/main/java/com/rackspace/salus/event/manage/model/kapacitor/Task.java
@@ -1,50 +1,0 @@
-/*
- * Copyright 2020 Rackspace US, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.rackspace.salus.event.manage.model.kapacitor;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import java.util.List;
-import java.util.Map;
-import lombok.Data;
-
-@Data
-public class Task {
-  String id;
-  Type type;
-  List<DbRp> dbrps;
-  String script;
-  Map<String, Var> vars;
-  Status status;
-  String error;
-  Stats stats;
-
-  public enum Type {
-    stream
-  }
-
-  public enum Status {
-    enabled, disabled
-  }
-
-  @Data
-  public static class Stats {
-    @JsonProperty("task-stats")
-    Map<String, Integer> taskStats;
-    @JsonProperty("node-stats")
-    Map<String, Map<String,Integer>> nodeStats;
-  }
-}
