+++ src/main/java/com/rackspace/salus/event/manage/services/TasksService.java
@@ -22,10 +22,10 @@ import com.rackspace.salus.event.discovery.EventEnginePicker;
 import com.rackspace.salus.event.manage.errors.BackendException;
 import com.rackspace.salus.event.manage.errors.NotFoundException;
 import com.rackspace.salus.event.manage.model.CreateTask;
-import com.rackspace.salus.event.manage.model.kapacitor.DbRp;
-import com.rackspace.salus.event.manage.model.kapacitor.Task;
-import com.rackspace.salus.event.manage.model.kapacitor.Task.Status;
-import com.rackspace.salus.event.manage.model.kapacitor.Task.Type;
+import com.rackspace.salus.event.model.kapacitor.DbRp;
+import com.rackspace.salus.event.model.kapacitor.Task;
+import com.rackspace.salus.event.model.kapacitor.Task.Status;
+import com.rackspace.salus.event.model.kapacitor.Task.Type;
 import com.rackspace.salus.event.manage.services.KapacitorTaskIdGenerator.KapacitorTaskId;
 import com.rackspace.salus.telemetry.entities.EventEngineTask;
 import com.rackspace.salus.telemetry.repositories.EventEngineTaskRepository;
