+++ src/main/java/com/rackspace/salus/event/manage/services/TestEventTaskService.java
@@ -27,12 +27,12 @@ import com.rackspace.salus.event.manage.errors.TestTimedOutException;
 import com.rackspace.salus.event.manage.model.TestTaskRequest;
 import com.rackspace.salus.event.manage.model.TestTaskResult;
 import com.rackspace.salus.event.manage.model.TestTaskResult.EventResult;
-import com.rackspace.salus.event.manage.model.kapacitor.DbRp;
-import com.rackspace.salus.event.manage.model.kapacitor.KapacitorEvent;
-import com.rackspace.salus.event.manage.model.kapacitor.Task;
-import com.rackspace.salus.event.manage.model.kapacitor.Task.Stats;
-import com.rackspace.salus.event.manage.model.kapacitor.Task.Status;
-import com.rackspace.salus.event.manage.model.kapacitor.Task.Type;
+import com.rackspace.salus.event.model.kapacitor.DbRp;
+import com.rackspace.salus.event.model.kapacitor.KapacitorEvent;
+import com.rackspace.salus.event.model.kapacitor.Task;
+import com.rackspace.salus.event.model.kapacitor.Task.Stats;
+import com.rackspace.salus.event.model.kapacitor.Task.Status;
+import com.rackspace.salus.event.model.kapacitor.Task.Type;
 import com.rackspace.salus.telemetry.entities.EventEngineTaskParameters;
 import com.rackspace.salus.telemetry.model.SimpleNameTagValueMetric;
 import io.micrometer.core.instrument.Counter;
