+++ src/test/java/com/rackspace/salus/event/manage/web/controller/TasksApiControllerTest.java
@@ -45,9 +45,9 @@ import com.rackspace.salus.event.manage.model.CreateTask;
 import com.rackspace.salus.event.manage.model.TestTaskRequest;
 import com.rackspace.salus.event.manage.model.TestTaskResult;
 import com.rackspace.salus.event.manage.model.TestTaskResult.EventResult;
-import com.rackspace.salus.event.manage.model.kapacitor.KapacitorEvent.EventData;
-import com.rackspace.salus.event.manage.model.kapacitor.KapacitorEvent.SeriesItem;
-import com.rackspace.salus.event.manage.model.kapacitor.Task.Stats;
+import com.rackspace.salus.event.model.kapacitor.KapacitorEvent.EventData;
+import com.rackspace.salus.event.model.kapacitor.KapacitorEvent.SeriesItem;
+import com.rackspace.salus.event.model.kapacitor.Task.Stats;
 import com.rackspace.salus.event.manage.services.TasksService;
 import com.rackspace.salus.event.manage.services.TestEventTaskService;
 import com.rackspace.salus.telemetry.entities.EventEngineTask;
